refactor multiprocessing to execute method, persist Stock Crawler
