+++ run.py
@@ -13,8 +13,6 @@ if __name__ == "__main__":
     print("--------------------\n" + \
           "Starting app...")
 
-    subprocess.run(["python", "crawler/stock_crawler.py"])
-    subprocess.run(["python", "stock_graphql_api/manage.py", "loaddata", "resources/output.json"])
     subprocess.Popen(["python", "stock_graphql_api/manage.py", "runserver"])
 
     print("--------------------\n" + \
