--- test.py
@@ -1,6 +1,0 @@
-import multiprocessing
-from crawler.stock_crawler import *
-
-
-if __name__ == "__main__":
-    StockCrawler()
